refactor: infra 도메인을 storage로 명칭 변경
자세한 사항은 issue #86 참고
