+++ src/main/java/com/umc/product/storage/domain/enums/StorageProvider.java
@@ -1,4 +1,4 @@
-package com.umc.product.infra.domain.enums;
+package com.umc.product.storage.domain.enums;
 
 public enum StorageProvider {
     AWS_S3,
